add support for creating a savestate from blip_buf.
the need for state support was required in my emulator (TotalSMS) runahead feature.
as blip_buf wasn't stored in the savestate, this resulted in discontinuities in the audio
when loading state (aliasing, pops etc).

blip_buf thankfully is very simple to dump the full context. blip_buffer however is a bit
more involved. as i don't use it in my emultor, i haven't bothered to impl it. If anyone needs this,
please email me.